+++ tests/integration/command__select_all_test.py
@@ -8,7 +8,6 @@ from tests.integration.helpers.givenpage import GivenPage
 
 def test_select_all_allows_to_overwrite_text_via_type(session_browser):
     browser = session_browser.with_(timeout=1)
-    browser.config.hold_driver_at_exit = True
     page = GivenPage(browser.driver)
     page.opened_with_body(
         '''